+++ GyverBraid/index.html
@@ -10,7 +10,10 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.6.0/p5.js"></script>
     <script src="https://cdn.rawgit.com/bitcraftlab/p5.gui/master/libraries/p5.gui.js"></script>
     <script src="https://cdn.rawgit.com/bit101/quicksettings/master/quicksettings.js"></script>
-    <script src="main.js?v=1.3.2" type="text/javascript"></script>
+    <script type="module">
+    	import * as gyver from './main.js?v=1.3.2';
+    	for (let nm in gyver) window[nm] = gyver[nm];
+    </script>
     <style>
         body {
             padding: 0;
